Revert Folder Move for Player Audio classes
